Introduce KeplerTable class
