+++ test/node/utils/layer-utils-test.js
@@ -20,7 +20,8 @@
 
 import test from 'tape';
 import {findDefaultLayer} from 'utils/layer-utils/layer-utils';
-import {findPointFieldPairs, createNewDataEntry} from 'utils/dataset-utils';
+import {createNewDataEntry} from 'utils/dataset-utils';
+import {findPointFieldPairs} from 'utils/table-utils/kepler-table';
 import {processCsvData, processGeojson} from 'processors/data-processor';
 import {GEOJSON_FIELDS} from 'constants/default-settings';
 import {LayerClasses, KeplerGlLayers} from 'layers';
